backend/requirements.txt
# AutoOps AI — Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
bcrypt==4.2.1
pyjwt==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sse-starlette==2.2.1
pandas
httpx==0.28.1
openpyxl

# Core engine dependencies (also needed)
openai==1.58.1
chromadb==0.5.23
sentence-transformers
numpy
